+++ src/solver/Constraint/LittleKillerConstraint.ts
@@ -1,12 +1,12 @@
 import { Board } from '../Board';
 import ConstraintBuilder from '../ConstraintBuilder';
 import { cellIndexFromName } from '../SolveUtility';
-import { ConstraintV2 } from './ConstraintV2';
+import { Constraint } from './Constraint';
 import { FPuzzlesLittleKillerSumEntry } from './FPuzzlesInterfaces';
 import { FixedSumConstraint, FixedSumConstraintParams } from './FixedSumConstraint';
 
 export function register(constraintBuilder: ConstraintBuilder) {
-    constraintBuilder.registerConstraint('littlekillersum', (board: Board, params: FPuzzlesLittleKillerSumEntry): ConstraintV2 | ConstraintV2[] => {
+    constraintBuilder.registerConstraint('littlekillersum', (board: Board, params: FPuzzlesLittleKillerSumEntry): Constraint | Constraint[] => {
         if (!params.value) {
             return [];
         }
